+++ src/readInput/readInput.ts
@@ -18,26 +18,31 @@ export function readInput(
   const { stdin, stdout } = process;
   // 获取一个随机字母
   const _n = () => getRandomInt(96, 122);
-  /** 获取唯一的 key  */
+  /** 获取唯一的 key
+   *
+   * 用于向数据仓储中添加本次调用的 key
+   */
   const uniKey = Symbol(String.fromCodePoint(_n(), _n(), _n()));
+
   return new Promise(resolve => {
     let rl: Interface;
     /**  注册事件
      *
      * 并在注册时指定当前是否开始或者是结束
+     *
+     * 将 process.stdout.keypress  事件放到回调中执行,然后再合适的时候再注销掉该事件
+     *
      */
     readInputData.on(uniKey, (runMi: boolean) => {
+      /// 意外退出（用于在 hot 中使用）
+      //  TODO  上面那么做会造成不必要的性能损耗,这里不做维护  -- (09:19:15-07/17/2024)
+      process.on('beforeExit', stdRemoveListener);
       /** 如果事件触发为 false ，则停止流的推入。等待事件的返回 */
       if (runMi) {
         rl = createInterface({ input: stdin, output: stdout });
         stdin.on('keypress', pressCallFn);
       }
     });
-
-    /// 意外退出（用于在 hot 中使用）
-    process.on('beforeExit', () =>
-      stdin.removeListener('keypress', removeListenerCallFn),
-    );
 
     /** Keyboard press callback
      *
@@ -46,7 +51,11 @@ export function readInput(
       // 如果当前并非第一个注册的方法先返回等待上一个注册的方法结束先
       /// 这里为了给列表做一个
       if (typeOf(_callback) == 'function') {
-        // 回调返回的是  true
+        /**
+         *  回调返回的是  true
+         *
+         *  则说明该方法已经结束，可以申请结束当前的移除监听工作
+         */
         if (Reflect.apply(_callback, null, [keyValue, key])) {
           removeListenerCallFn();
           resolve(key);
@@ -63,12 +72,22 @@ export function readInput(
      *
      * 移除监听 */
     function removeListenerCallFn() {
+      //  TODO  当前这么做或导致每一次都不停的创建监听,移除监听.后续可能会优化  -- (09:40:01-07/17/2024)
+      process.removeListener('beforeExit', stdRemoveListener);
+      stdRemoveListener();
       // 调用移除将使用权交给先一个使用者或者结束 read input
-      stdin.removeListener('keypress', pressCallFn);
       rl.close && rl.close();
       // 移除并调用下一个
       readInputData.remove;
       return true;
+    }
+
+    /***
+     *
+     *
+     */
+    function stdRemoveListener() {
+      stdin.removeListener('keypress', pressCallFn);
     }
   });
 }