+++ test/testCursor.ts
@@ -1,11 +1,8 @@
 import { test } from 'node:test';
 
 /** 测试光标的位置 */
-test.todo('test', async () => {
+test.skip('test', async () => {
   // readInput(() => {
   //     return true
   // });
 });
-
-/** 测试光标部分 */
-export default () => 'test cursor';